requirements.txt
Pillow
kivy
kivymd
